Topics controller and views updated
The topic index view has been updated to list the topic name. The Index
method of the TopicsController populates a list of topics, each with a
list of posts.
Still need to get the count of posts somehow for displaying on the index
page. It would be great to have a property which states datetime of last
post made in the topic.